# docker_data/redoc/swagger.yaml
openapi: 3.0.0
info:
  title: Geostories API
  version: 1.0.0
  description: API for managing Geostories users
servers:
  - url: http://localhost:3000/
paths:
...
  /tours:
    get:
      summary: Get all tours
      security:
        - bearerAuth: []
      tags:
        - Tours
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tour"
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
  /tours/{id}:
    get:
      summary: Get a tour by ID
      security:
        - bearerAuth: []
      tags:
        - Tours
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the tour to retrieve
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
  /tours/started/:
    get:
      summary: Get all started tours for the authenticated user
      tags:
        - Tours
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StartedTour"
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
  /tours/started/{id}:
    get:
      summary: Get started tour by ID
      security:
        - bearerAuth: []
      tags:
        - Tours
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the started tour to retrieve
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartedTour"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
  /tours/start:
    post:
      summary: Start a new tour
      security:
        - bearerAuth: []
      tags:
        - Tours
      requestBody:
        description: The tour to start
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tourId:
                  type: string
                  description: ID of the tour to start
              required:
                - tourId
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartedTour"
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
  /tours/stop:
    put:
      summary: Stop a started tour
      security:
        - bearerAuth: []
      tags:
        - Tours
      requestBody:
        description: The started tour to stop
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                startedTourId:
                  type: string
                  description: ID of the started tour to stop
              required:
                - startedTourId
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartedTour"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
...
  /tour/path/name/{continentName}/{countryName}/:
    get:
      summary: Get tours of a country
      security:
        - bearerAuth: []
      tags:
        - Tours
      parameters:
        - in: path
          name: countryName
          schema:
            type: string
          required: true
          description: country name of tour
        - in: path
          name: cityName
          schema:
            type: string
          required: true
          description: city name of tour
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tour"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
...
  /tour/path/id/{continentId}/{countryId}/:
    get:
      summary: Get tours of a country
      security:
        - bearerAuth: []
      tags:
        - Tours
      parameters:
        - in: path
          name: continentId
          schema:
            type: string
          required: true
          description: continent id of tour
        - in: path
          name: countryId
          schema:
            type: string
          required: true
          description: country id of tour
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tour"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StartedTour:
      type: object
      properties:
        id:
          type: string
        isCompleted:
          type: boolean
        dateStarted:
          type: string
          format: date-time
        dateEnded:
          type: string
          format: date-time
          nullable: true
        tourId:
          type: string
        userId:
          type: string
...
    TourPoint:
      type: object
      properties:
        id:
          type: string
        longitude:
          type: number
        latitude:
          type: number
        description:
          type: string
        dateCreated:
          type: string
          format: date-time
        dateUpdated:
          type: string
          format: date-time
        tourId:
          type: string
          nullable: true
    Tour:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        image:
          $ref: "#/components/schemas/Image"
        tourPoints:
          type: array
          $ref: "#/components/schemas/TourPoint"
        dateCreated:
          type: string
          format: date-time
        dateUpdated:
          type: string
          format: date-time
        xp:
          type: number
        cityId:
          type: string
          nullable: true
...
    Image:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        name:
          type: string
        dateCreated:
          type: string
          format: date-time
        dateUpdated:
          type: string
          format: date-time

    #* --- With includes ---
...